+++ AoE2ScenarioParser/objects/data_objects/terrain_tile.py
@@ -3,7 +3,7 @@ from typing import Optional, Tuple
 from AoE2ScenarioParser.datasets.terrains import TerrainId
 from AoE2ScenarioParser.helper.helper import i_to_xy
 from AoE2ScenarioParser.objects.aoe2_object import AoE2Object
-from AoE2ScenarioParser.scenarios import scenario_store
+from AoE2ScenarioParser.scenarios.scenario_store import getters
 from AoE2ScenarioParser.sections.retrievers.retriever_object_link import RetrieverObjectLink
 
 
@@ -42,7 +42,7 @@ class TerrainTile(AoE2Object):
     @property
     def xy(self) -> Tuple[int, int]:
         if not self._xy:
-            self._xy = i_to_xy(self._index, scenario_store.get_map_size(self._host_uuid))
+            self._xy = i_to_xy(self._index, getters.get_map_size(self._host_uuid))
         return self._xy