+++ AoE2ScenarioParser/objects/managers/de/xs_manager_de.py
@@ -4,7 +4,7 @@ from typing import Union
 from AoE2ScenarioParser.helper.printers import warn
 from AoE2ScenarioParser.objects.aoe2_object import AoE2Object
 from AoE2ScenarioParser.objects.data_objects.trigger import Trigger
-from AoE2ScenarioParser.scenarios import scenario_store
+from AoE2ScenarioParser.scenarios.scenario_store import getters
 from AoE2ScenarioParser.sections.retrievers.retriever_object_link import RetrieverObjectLink
 from AoE2ScenarioParser.sections.retrievers.support import Support
 
@@ -67,7 +67,7 @@ class XsManagerDE(AoE2Object):
         )
         self.xs_trigger.new_effect.script_call(message="")
 
-        trigger_manager = scenario_store.get_trigger_manager(self._host_uuid)
+        trigger_manager = getters.get_trigger_manager(self._host_uuid)
         trigger_manager.import_triggers([self.xs_trigger], insert_index)
 
     def _append_to_xs(self, title, string) -> None: