+++ AoE2ScenarioParser/scenarios/scenario_store/getters.py
@@ -1,40 +1,11 @@
-from typing import TYPE_CHECKING, Dict, Optional, List, Union, Type, Tuple
+from typing import Optional, List, Tuple, Dict, TYPE_CHECKING
+from AoE2ScenarioParser.scenarios.scenario_store import store
 
 if TYPE_CHECKING:
     from AoE2ScenarioParser.objects.data_objects.unit import Unit
     from AoE2ScenarioParser.objects.managers.trigger_manager import TriggerManager
-    from AoE2ScenarioParser.scenarios.aoe2_scenario import AoE2Scenario
     from AoE2ScenarioParser.scenarios.aoe2_de_scenario import AoE2DEScenario
     from AoE2ScenarioParser.sections.aoe2_file_section import AoE2FileSection
-
-_scenarios: Dict[str, 'AoE2Scenario'] = {}
-
-
-def _get_scenario(uuid: str) -> Optional[Union['AoE2Scenario', Type['AoE2Scenario']]]:
-    """
-    Get scenario through uuid. Not intended to be called outside of the store itself.
-
-    Args:
-        uuid (str): The UUID of the scenario
-
-    Returns:
-        The scenario based on it's uuid
-    """
-    if uuid == "<<NO_HOST_UUID>>":
-        return None
-    return _scenarios[uuid]
-
-
-def register_scenario(scenario: 'AoE2Scenario') -> None:
-    """
-    Register a scenario to the store
-
-    Args:
-        scenario (AoE2DEScenario): The scenario to register
-    """
-    if scenario.uuid in _scenarios:
-        raise ValueError("Scenario with that UUID already present")
-    _scenarios[scenario.uuid] = scenario
 
 
 def get_unit(uuid: str, unit_reference_id: int) -> Optional['Unit']:
@@ -66,7 +37,7 @@ def get_units(uuid: str, unit_reference_ids: List[int]) -> Optional[Tuple[List['
         A tuple with a list of the found unit objects and a list of the IDs that weren't found.
     """
     unit_reference_ids = unit_reference_ids.copy()
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario:
         result = []
         for unit in scenario.unit_manager.get_all_units():
@@ -87,7 +58,7 @@ def get_sections(uuid: str) -> Optional[Dict[str, 'AoE2FileSection']]:
     Returns:
         The sections of the selected scenario
     """
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario:
         return scenario.sections
     return None
@@ -103,7 +74,7 @@ def get_scenario_version(uuid: str) -> Optional[str]:
     Returns:
         The scenario version of the selected scenario (e.g. '1.43')
     """
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario:
         return scenario.scenario_version
     return None
@@ -119,7 +90,7 @@ def get_game_version(uuid: str) -> Optional[str]:
     Returns:
         The game version of the selected scenario (e.g. 'DE')
     """
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario:
         return scenario.game_version
     return None
@@ -135,7 +106,7 @@ def get_map_size(uuid: str) -> Optional[int]:
     Returns:
         The map size of the scenario
     """
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario:
         return scenario.map_manager.map_size
     return None
@@ -152,7 +123,7 @@ def get_trigger_name(uuid: str, trigger_index: int) -> Optional[str]:
     Returns:
         The name of the trigger with the given ID
     """
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario and trigger_index < len(scenario.trigger_manager.triggers):
         return scenario.trigger_manager.triggers[trigger_index].name
     return None
@@ -169,7 +140,7 @@ def get_variable_name(uuid: str, variable_index: int) -> Optional[str]:
     Returns:
         The name of the variable with the given ID
     """
-    scenario: 'AoE2DEScenario' = _get_scenario(uuid)
+    scenario: Optional[AoE2DEScenario] = store.get_scenario(uuid)
     if scenario:
         if gv := get_game_version(uuid) == "DE":
             variable = scenario.trigger_manager.get_variable(variable_index)
@@ -179,6 +150,22 @@ def get_variable_name(uuid: str, variable_index: int) -> Optional[str]:
                 return f"Variable {variable_index}"
         else:
             raise ValueError(f"Scenarios with the game version: {gv} do not support variables.")
+    return None
+
+
+def get_trigger_version(uuid: str) -> Optional[float]:
+    """
+    Get the trigger version of the scenario.
+
+    Args:
+        uuid (str): The UUID of the scenario
+
+    Returns:
+        The trigger version.
+    """
+    scenario = store.get_scenario(uuid)
+    if scenario:
+        return scenario.sections['Triggers'].trigger_version
     return None
 
 
@@ -192,7 +179,7 @@ def get_trigger_manager(uuid: str) -> Optional['TriggerManager']:
     Returns:
         The trigger manager of a scenario.
     """
-    scenario = _get_scenario(uuid)
+    scenario = store.get_scenario(uuid)
     if scenario:
         return scenario.trigger_manager
     return None