+++ AoE2ScenarioParser/sections/dependencies/dependency_eval.py
@@ -1,5 +1,5 @@
 class DependencyEval:
-    def __init__(self, eval_code, eval_locals=None, values_as_variable=None):
+    def __init__(self, eval_code, eval_locals=None):
         """
         Object for storing dependency eval code and it's locals.
 
@@ -9,11 +9,9 @@ class DependencyEval:
         """
         if eval_locals is None:
             eval_locals = {}
-        if values_as_variable is None:
-            values_as_variable = []
+
         self.eval_code = eval_code
         self.eval_locals = eval_locals
-        self.values_as_variable = values_as_variable
 
     @classmethod
     def instance_or_none(cls, eval_code):